block: add missing block_unlock() in block_dev_sync_block_async()
Pretty simple, in the case where `block_dev_sync_block_async()` was
called on a block that was already valid and not dirty, it would skip
the sync but leave the block locked, effectively leaving it locked
forever and locking up parts of the system.

The check is still valid, but we need to unlock the block if we're not
going to call into the block-dev sync routine which would have normally
done the unlock for us.

Interestingly, this is a situation which typically doesn't happen
because these flags typically get checked before calling sync anyway,
but it can manifest itself if you do multiple bread()'s of the same
block very quickly. All of them attempt to sync the block (due to
argubily incorrect logic) and on the second sync when the block is valid
and not yet dirty the lock is never released after sync is called.