check-ext2: test multiple different disk configurations
Previously we only ran the ext2 tests with  a single default disk. This
change allows us to define scripts to create test disks with custom
settings, and then run the ext2 tests with every test disk we specify.

This is necessary for testing some file-system wide settings, such as
the inode size.