ext2: support non-128-bit inode sizes
ext2 has a super-block setting that describes the size of the inode's on
disk. This setting has to be a power of two, and small then the block
size, but it can be larger than the 128-bit disk inode we use (I assume
this is for storing extended attributs, which we don't support).