mm: add address_space_find_region()
This function finds a hole of pages of a paticular size in the provided
address_space. It's a bit crude, but it works well enough for the
moment. We also have a minimum address, where we try to find a spot
above that location. The idea is that we want to avoid getting in the
way of the brk if that is used for allocation.