block: replaced block_mutex with BLOCK_LOCKED flag
This overall simplifies some of our logic. The block_mutex is gone, and
instead we have the BLOCK_LOCKED flag. We're also now protecting the
flags with a spinlock, and have a wait_queue for flag changes.

When trying to lock the block, you take the spinlock, and then wait on
the wait-queue if BLOCK_LOCKED is already set.