event: add event protocol
The event protocol is a simple way for userspace to read from a stream
of "events" from the kernel. Each event is a very simple 8-byte value,
indicating the kind of event, and also a code and value telling what
happened.

The event_queue structure implements most of the logic related to the
event protocol. To impement something using it, you simply write your
own open() that calls event_queue_open(), and then call
event_queue_submit_event() when you have a new event (This can be called
from any context).