keyboard: allow turning off keyboard input to TTY
With the new keyboard event interface, a program may want to use the
event interface and stop keyboard events from going to the active TTY.
To achieve, this we have a new keyboard_set_state() call which turns
keyboard input to the TTY on and off.

You can also now press the Print Screen button to force keyboard input
to go back to the TTY, in the event a program left it off by mistake.