screen: pass screen pointer to callbacks, add refresh()
This is preparation for adding new screen implementions. Some
implementations will require access to the `struct screen` object in the
screen callbacks. Others will also require an explicit `refresh()` call
to indicate the screen should be redrawn, which has also been added in a few
spots.

We also added a function that lets you swap the currently active
`struct screen`.