fbcon: framebuffer based console implementation
The fbcon is a screen implementation that is rendered onto a backing
framebuffer supplied by a driver or other means. The driver fills in the
`fb_info` structure and then feeds it to fbcon to start the console.

The framebuffer is exposed via a device which userspace can open. It
includes a few ioctls for getting dimensions and information, one for
turing off the console rendering (for programs that want exclusive
access), and one for mapping the framebuffer memory into the process
calling it. We do not yet have a proper userspace `mmap`, so we use an
ioctl() for this.