x86: add multiboot2 support
We currently support multiboot 1, and want to continue duing that as
that's what qemu supports for the `-kernel` parameter. That said,
when booting via GRUB multiboot2 offers us more in terms of what we can
ask from the bootloader, and in paticular GRUB can provide us with a
framebuffer when using multiboot2, which will greatly simplify the
process of getting one on a real PC.

Thus, we support both. In the cmain function we check the magic to
determine which one was used to boot us, and then parse the multiboot
header for the information we need (Right now just memory and
command-line information).