block: replace mutex with spinlock, clarify locking
The block_cache.lock was replaced with a spinlock. It still holds the
same role and is much more efficent for something that gets hit a lot.
Generally speaking, most of the code could stay the same. A few
operations were moved outside of the lock, though, in paticular:

1. The lock is dropped before calling block_new(). Duplicates created
   due to races are deleted.

2. During syncing, we no longer drop and reacquire the lock, instead we
   place all the blocks directly on the sync_list and then sync them all
   (without or without waiting) in one go after dropping the lock. This
   also has the advantage that we won't miss any blocks adding during
   the time the lock was previously dropped.

In addition, the locking around the block flags was clarified and
slightly fixed. Modifying the flags always requires taking flags_lock,
and looking at them also requires taking flags_lock. There is a special
case for DIRTY and VALID though, which is that they can be checked
*without* the lock if the block is LOCKED and you are the one who LOCKED
it. This allows most accesses in the block driver code to be lock-free.

Fixes: #46