ktimer: add timer_cancel()
timer_cancel() is like timer_del() but it also ensures the removed timer
is not currently running, where as timer_del() simply tells you whether
or not it got scheduled. We do this by storing a pointer to the
currently running timer, and do a direct pointer comparison to detect if
the canceled timer is the currently running one - simple, but effective.
The waiting is just a busy wait, but timer should exit relatively fast
so this shouldn't be that big of a deal.

Note that, without an SMP machine this is impossible to trigger. Timers
run with interrupts off, so there's no way for other code to run while a
timer is the currently running timer.