wait: fix ordering of signal check
Not sure what the logic was on this one. We have to check for pending
signals *after* setting the sleep condition so that there's no race.

The signal logic will wake up our task and set us to running. This
avoids a race because if a signal comes in *after* we check for signals,
but *before* we actually yield, our state will be set to running and
we'll simply get scheduled and run again.

However, by checking the state of the signals *before* we set our task
state to sleeping, there's a small window after we check for signals but
before we set the state to sleeping where we could receive a signal and
a wake-up. The issue would then be that immediately after that wake-up
we would set our state to sleeping and the wake-up would be lost.
Checking for signals after setting the state avoids this lost-wakeup
problem.