kernel: new API for sleeping for a set amount of time
Currently this is done by setting the task->wake_up field, but that API
is a bit crusty and is going away. This new API makes use of ktimer's to
do the sleeping, which is relatively simple. It supports both
interruptable and uninterruptable sleeping.

It also fixes a subtle bug where `sys_usleep()` was rounding down, resulting
in sleeping for less time than required.