mknod: fix handling of mode argument
When the logic for umask was added, we incorrectly started clearing out
the upper bits of the mode argument, which clears the inode type to zero
and breaks the creation of nodes like block-devices and char-devices.

Fixing this just requires correctly handling the format bits - remove
them, check that they are valid, and then add them back after applying
the umask.