kernel: change dev_t to be 32-bit with a special format
Previously, I effectively copied the format of dev_t from the Linux
kernel, including copying the legacy detail of userspace dev_t being
16-bit, which causes a fairly small limit on the number of devices.

But there's really no reason to actually keep this, and changing to a
32-bit format makes it much easier to support more partitions per
device and more devices overall. It also allows us to match the "kernel"
format to the "userspace" format and remove some of the weird things on
on there.

The ending format I went for is 12-bits for major and 20-bits for minor.
This matches what the Linux Kernel uses internally (more or less). The
biggest reason for going for this format is simply that it's what gets
stored in `ext2`, which gives us better support for it. If we went for
an even split of 16/16, we couldn't store the every possible value in
`ext2` and have the major/minor numbers still correct. And having more
minors means more partitions, so in general the 12/20 split makes sense.