block: add new block-device interface
The new block device interface is based around the `struct disk`. It
represents a cachable device that can have blocked read from and written
to it.

The `struct block_device` format is changed substantially, and we now
allocate a `struct block_device` *per* dev_t, rather than have one-per
major. This simplifies a lot of the logic because it gives every dev_t a
separate list of blocks and block_size, etc.

When new disks are detected on the system, a new `struct disk` is
allocated and filled out with the proper device number information, and
then registered with the system via `disk_register()`. Once this is
done, they can be accessed by `disk_get()` or more commonly via grabbing
the `struct block_device` associated with that disk.

When a paticular `struct block_device` is opened, it automatically looks
up the coresponding `struct disk` (return -ENXIO if it doesn't exist),
opens it, and configure it if necessary (IE. Sets up the partition
table). It then sends blocks off to the disk for syncing when
necessary, and offers the existing `block_get()` `block_put()`
interface.

Note that with this change, more of these structures are now reference
counted. In paticular `struct block_device` needs to have the reference
to it dropped in the proper places.

Fixes: #34