inode: drop block_device reference if the inode has one
If an inode is a block device, then it holds a reference to the backing
`struct block_device`, which needs to be dropped if/when the inode is
free'd. We can otherwise keep it around as long as we want.

Note that we need to chagne some of our logic because we don't want to
call `block_device_put()` while holding the inode cache spinlock. This
just involves splitting up the process of removing the inode from the
cache hashtable and actually freeing the inode.