ext2: respect the inodes_per_block_group setting
Ext2 has a setting that defines the number of inodes per block group.
Typically, this is just the number of bits in a block, to match the
largest size the bitmap can be. It can, however, be smaller than that,
and we shouldn't allocate inodes past this value in a group.