ext2: truncate: handle vfs_bmap() returning SECTOR_INVALID
During a truncate that increases the size of an inode, the new space is
supposed to be cleared out with zeros. This *generally* works without
any work because we only allocate the blocks when asked during a
`bmap_alloc()` call, and at that point we zero the new blocks.

That said, we still need to handle the case where the size was not on a
block boundary and was expanded. In that situation, if the last block in
the file was already allocated, we need to clear the remaining space
with zeros. We were kinda doing this before, but not in the expansion
case (the important one), and we were also not verifying that
`vfs_bmap()` returning a valid sector - if that block was never touched,
it may not exist.