ext2: bmap: correctly handle out-of-space situations
Durring the process of mapping a new block into a file, it may require
allocating one or more new blocks during that time. All of those
allocations can fail if the drive is out of space, and we should
correctly handle those cases and propagate that error back to the user.

The nice detail is that we are already mapping the blocks in the
"correct" way starting from the blocks directly referenced by the inode,
and then allocate blocks in order for the blocks one, two, or three hops
away. This is important because it means if an out-of-space situation
happens half-way through mapping a triple-indirect block, we don't need
to release *every* block, just the one we were trying to map. The
indirect, doubly indirect, and tripply indirect blocks are all correctly
mapped to the inode and thus will get correctly released during a
truncate operation.