tests: ext2: increase memory on test VMs
The qemu VMs for testing the ext2 FS driver only get 128MB of memory by
default, and with some of our larger tests like the out-of-space test we
can accidentally run out of memory and stall the kernel.

A long-term fix is to address the OOM issues so that the kernel can
clear some of the internal caches and continue in these situations, but
bumping up the memory to 512MB avoids the issues for now.