bcache: expose block_get_nosync() externally
block_get_nosync() is useful in the rest of the kernel because there are
very common situations where we want to create a new block but we don't
care about the disk contents (Because Ex. We're just going to zero it).
In those situations, it's a good optimization to simply clear out the
block and call `block_mark_synced()` without actually touching the disk.
That's easy enough for callers to do if we simply expose
`block_get_nosync()` to them.