ata: add basic DMA support
This was removed when the ATA driver was rewritten. This does not yet
support request grouping into a single DMA request, but it does support
transfering blocks via DMA rather than PIO. DMA support is checked for
via BAR4 of the PCI device along with checking the results of the
IDENTIFY command.

Fixes: #39