block: base max cache size on bytes, not blocks
Previously, the cache size was limited by number of blocks, which mean
the cache could be 4 or 8 times smaller if a smaller block-size is used,
which is not ideal. This changes the block-cache size to be based on the
number of bytes the blocks take up instead, which allows the same amount
of space to be taken up by any number of different size'd blocks.