tests: ext2: bump time limit to 3 minutes
The time the tests take when running on GitHub are a bit all over the
place, and the block-size-1024 tests (the slowest ones) can occasionally
bump around 90 seconds, and in some rare cases hit the 120 second time
limit. For the moment the easiest fix is to just move to 3 minutes,
which should be long enough.