kp: Only render the long string once, call outputs without lock
This is a fair fundimental change to how kp() logging works. Previously,
the actual va_list was passed to each individual kp_output, where it
would then render it with it's own printf_backbone. Also, due to the way
the list of kp_output's worked, all of these functions were called under
a spinlock.

Now, the kp() line is only rendered once (to a string) and then that
string is provided to each kp_output. In the future, we'll likely only
store it to the `klog` and then have the other outputs read from `klog`,
but this is a start.

Also in this, the kp_output's are now reference counted and the spinlock
is dropped while calling their `print` function, allowing them a bit
more freedom in what they do (They still can't take mutex's and stuff
because they may be called from an interrupt context). This reference
counting logic allows them to be safely added and removed at runtime.

Each output also gets its own log-level, in addition to the global max
log level, and all of them can be changed at runtime.

Fixes: #27