creds: protect credentials with a spinlock
Currently this uses a mutex, which is both wastefull and also introduces
a nasty bug in the task API where we try to take the creds mutex while
holding the ktasks spinlock.

Only one piece of code had to be modified to allow the creds to be
protected with a spinlock, because the code called
user_copy_from_kernel_indexed() (which can sleep) while holding the
lock. The simple solution is to simply make a copy of the groups and
then drop the lock before copying them to userspace.

Fixes: #53