kparam: add automatically set kernel parameters
This change allows defining kernel parameters via the KPARAM() macro,
and at boot all the kparam's will automatically be parsed and set
if they were provided.

The design comes from the Linux kernel, and is mostly just a bit of
linking magic like ktest. `struct kparam` describes a kernel parameter,
and all of them are placed into the `.kparam` section. At boot, we loop
over all the entries in the `.kparam` section, and if a kernel parameter
matching the name was provided, the value from that parameter is
assigned to the provided lvalue. A simple type enum (and static assert
at compile time) indicate the type of parameter.

Fixes: #57