kparam: integrate cmdline and kparam together
This removes the existing 'kernel_cmdline_*' API in favor of always
using the kparam API.

The reasoning is pretty simple - kparam can handle an arbitrary number
of commandline parameters, but 'kernel_cmdline_*' has to store them
separately and thus introduces a hard limit.

Because when using kparam you're just going to store that value
somewhere-else anyway, there's no reason to retain all the cmdline
parameters in an array that's never actually going to get used. So we're
removing the existing 'kernel_cmdline_*' API and we simply resolve the
kparam entries as we parse the commandline.