kp: add separate max-log-level wrappers
kp_check_level and kp_force are used to create kp wrappers that use a
separate max-log-level from the global one. This will allow more
specific kp wrappers which can be enabled or disabled separate from
everything else. That allows us to do things like enabling debug or
trace for a specific part of the kernel without enabling it for
everything else.