slab: don't call palloc() while holding the slab spinlock
Currently, the slab code is bugged as it keeps hold of the slab spinlock
while calling palloc(), and because the palloc() flags are provided to
us by the caller, the call to palloc() might attempt to sleep. This is
pretty simple to fix, because we don't even touch the slab structure
until after we have allocated the chunk of memory from palloc(), so
we're free to simply drop the lock during the call to palloc() and keep
everything the same.

The only small catch is that this creates a race where two different
callers can both allocate a new slab frame, and then both of the new
empty frames get added to the slab's list of frames. To combat this
slightly, the 'allocate' code will restart the loop over all the frames
after one is added, rather than simply allocate a entry from within the
newly-added frame as we were doing before.

In a race situation, there may actually now be new space in the frames
before the newly-added one (And in the race situation, this is basically
guaranteed since a new frame was just added before ours). Restarting the
'allocate' code ensures that any new free space is used before the new
frames are touched. This logic keeps the new frames empty unless we run
out of free space in all the other frames, and thus keeps the new frames
ripe for freeing if the OOM is called.

Fixes: #36