mm: add bootmem abstraction
bootmem is a simple bump allocator designed for early boot. It supports
a fixed number of memory regions and then allocates from the first one
with a large enough region inside (respecting the requested alignment).

It also supports the ability to automatically initialize palloc and
transfer the region information (without any allocated regions) to
palloc. This allows the arch code to simply register as many regions as
they want with bootmem, use bootmem as necessary, and then transition to
palloc when they are able without having to register regions with palloc
again.

bootmem also automatically excludes the memory that the kernel occupies,
simplifying its usage.