x86: replace kbrk with bootmem
This is fairly straight-forward - kbrk usages become bootmem usages
instead. This also removes the need to pass around the kbrk pointer,
which cleans up the code a bit.

The initialization code now seeds bootmem, which also allows it to gain
the ability to register more than one region (as bootmem can handle
that, but kbrk couldn't).