palloc: correctly check if addr is between kern_end and kern_start
palloc does a sanlity check on the pages allocated and free'd to ensure
they are not part of the kernel's memory (and complains loudly if they
are), as obviously that memory is in use and something must be bugged in
those situations.

Currently, the check simply checks if the address is less than kern_end,
which was sufficent when we only handled one memory region (which was
always after the kernel). However, now that the boot allocator is a bit
fancier and allows multiple regions, some of them may be before the
kernel in memory. Thus, palloc should not warn in that situation and
should instead correctly check that the address is *between* kern_start
and kern_end before warning.