char_dev: supply a default file ops for every character device
This is pretty simple and solve a potential panic when a char device
that doesn't exist is attempted to be opened. The starting fops for
character devices are now always the same, and only have an open()
method. When open() is called, it checks if the `cdev` field on the
inode is filled in, and if so replaces the fops with that cdev's one and
calls open() on that cdev. If the `cdev` field is not filled in, it
returns -ENODEV instead.