event: add device hotplug event stream
This adds a very simple event stream of device add/remove events. The
events include the type of device (block or char) and dev_t of the
device. This event stream will be used by a userspace daemon to
dynamically update the contents of `/dev`.

To make this function after this addition, the rest of the kernel needs
to be updated to call this function whenever they register a new char or
block device. For block devices, this is simple logic that can be
integrated into the `struct disk`. For char devices, because there isn't
really a unified abstraction around registering them calls need to get
sprinkled around more of the kernel.