block: disk: report disk add/remove events
This modifies the disk logic to call event_submit() to report disk
additions and removals. Generally this is pretty simple - disk register,
and partition additions trigger add events. Partition removal and disk
removal trigger remove events.

The logic is modified slightly to simplfy the reporting of the whole
disk vs. partitions. The "whole" disk is registered as a regular
partition, so by moving things around we can utilize that to trigger the
partitions and whole disk add event. Partition removal is a bit
different since we don't ever remove the whole disk, but it's still
pretty straight forward.