char: submit events on char device additions
This modifies all the existing character devices in the kernel to submit
a device addition event if their character device exists. In some cases,
this is unconditional, in other cases it is dependent upon the
underlying hardware being present.

Wiring this up also required wiring up a bunch of extra init functions.
Hopefully soon we'll have a better solution for those.