coreutils: add devd, userspace daemon for managing /dev
Currently, all the nodes in /dev need to be statically created before
the OS runs - this has obvious limitations, both from a scalability
standpoint and a discoverability standpoint. Nobody wants N number of
disk block devices and 256 partition devices for each, but you can
technically have that many. Much better is having a `/dev` that only
consists of devices that actually exist on the system.

To achieve this, there is a new devd daemon which reads from
/proc/devices to recieve device events from the kernel. On additions or
removals it automatically creates or removes the /dev entry for that
device, along with applying the proper permissions and name depending on
what it is.

Fixes: #44