init: add new initcall dependency system
This adds a new initcall implementation, similar to Linux's, which
allows defining initcalls without needing to tie them all together with
global functions like we do now. Instead you now define your function,
and then define an initcall with one of the `initcall_*` macros (for
specifying which part of the initialization sequence it should be a part
of). Extra dependencies can also be defined via the
`initcall_dependency()` macro. When the kernel is built, the initcalls
are sorted according to their dependencies, and then an array of all the
initcalls is created which will be processed at boot time.