coreutils: ping: fix timeval-to-ms calculation
The timeval returned by the kernel used to be bugged, which had the
effect of making the buggy time code in ping work correctly. But since
that was fixed in the kernel a while back, the ping code now converts
the timeval to ms incorrectly, which makes it think time has gone much
faster than it is. Specifically, the tv_usec value is now actually
useconds, rather than microseconds when it was bugged, so it needs a
'/ 1000' that wasn't necessary before.