check: user_check: limit stack size to silence warning
Since adding stack size limit warnings, these tests were giving off a
warning. `len` is a test input and is never large enough (in our tests)
to actually overflow the stack, but it still  triggers the warning. To
avoid the warning this change introduces a maximum size for the input.