coreutils: devd: reset ownership when recreating dev node
If for some reason an existing node has the wrong device number, but the
right owernship, then we'll incorrect recreate the device node with the
correct number, but fail to reset the owernship back to the correct
values (resulting in root/root).

It happens because we continue to use the old stat() info even after
doing an unlink() and mknod() that invalidate that information. The easy
solution is to simply set the 'err' value to force the chown() to
happen.