coreutils: devd: correctly create the devd log file
We were missing O_CREAT, which meant /var/log/devd.log would not get
created if it does not exist. We were also leaving stdout buffered,
which meant a lot of our printf's would not actually make it to the file
and would be left in the cache when devd is killed on power-off.