super: remove super_block from list on dealloc()
Typically this isn't an issue because the umount() logic already has
this list_del() in a proper place. However, the mount() logic has a few
failure states where a dealloc() happens when the super_block is already
in the list. In those cases, a list_del() needs to happen. dealloc() is
the simplest place to put it.