block: add loop block device
The loop block device is fairly straight-forward - it hold a reference
to an inode, and when a block r/w is requested for the loop device, the
sync_block() logic simply calls bmap() on the inode, reads the block,
and copies from one block to another. There's a bit of logic realted to
handing differing block-sizes on the loop device and the inode's block
device, but it's not too bad.

Somewhat interesting is how the block syncing is handled. We don't want
every sync_block() call to the loop device to block while waiting for
the backing block device, so instead we queue the block to be synced on
a workqueue in the background. This isn't quite as perfect as it could
be because the workqueue callbacks still block waiting for the
destination blocks to sync (The best solution would involve registering
a `struct work` with the wait_queue on destination blocks), but this
solution balances speed and complexity.

Fixes: #61