coreutils: mount: display current mounts when no arguments are passed
Fairly simple, when no arguments are given to 'mount' we now parse the
contents of /proc/mounts and print out the results. This contains very
similar logic to `df`, where when we recieve an entry that is a raw
device number ('(major,minor)'), we walk the device nodes in `/dev` to
find one with a matching number and list that as the device.

Fixes: #5