kmalloc: automatically service large allocations with palloc()
Currently, kmalloc() can only service allocations as large as it's
largest slab, and if we go over that it panics. That's not really ideal,
but we also don't want to add any larger slabs. The simple solution is
to simply have kmalloc() call into palloc() in the event of an
allocation that is too large. We then create simple descriptors that
list the address and number of pages that were allocated, and then call
pfree() when the address is free'd. This way, kmalloc() can be used for
larger generic allocations as well and it works just fine.

Note that the large allocation tracking is very unoptimal, but at the
moment we're going to have very few users so it shouldn't be that big of
a deal.