stddef: add scoped_using_assign() macro
scoped_using_assign() is a macro very similar to scoped_using_cond().
The main difference is that scoped_using_assign() is designed for
functions which *return* the thing you then want to clean up. For
example, block_get() and block_put() are ideal users, because
block_get() returns the block that then needs to get block_put(). This
is different from the cond version where it expects you to be performing
functions on something that already exists (like a mutex).