console: allow resizing the console
This is a little messy, but overall the solution was pretty simple. The
main thing that makes this possible is that we allocate the new console
screen buffers before taking all the vt locks. This allows us to do
normal allocations before we hold the spinlocks, and makes it easy to
get the new buffers. The rest is just a matter of swapping the buffers
and correctly transfering the old data, send the tty resize if
necessary, and do some cleanup. As a result of this change, the
locations of some of the static structures has changed (because the
arch-generic console doesn't know anything about the default arch screen
size), and all of the vt code has been changed to use the `rows` and
`cols` values rather than the constants.

Fixes: #35